Further refines timeline code
