+++ moped-editor/src/views/projects/projectView/ProjectPhases.js
@@ -4,14 +4,7 @@ import React from "react";
 import {
   Button,
   CircularProgress,
-  Grid,
-  TextField,
-  Switch,
-  Select,
-  MenuItem,
   Typography,
-  FormControl,
-  FormHelperText,
 } from "@material-ui/core";
 import {
   AddCircle as AddCircleIcon,
@@ -21,7 +14,6 @@ import MaterialTable, {
   MTableEditRow,
   MTableAction,
 } from "@material-table/core";
-import { handleKeyEvent } from "../../../utils/materialTableHelpers";
 import typography from "../../../theme/typography";
 
 // Query
@@ -38,6 +30,9 @@ import parseISO from "date-fns/parseISO";
 
 // Helpers
 import { phaseNameLookup } from "src/utils/timelineTableHelpers";
+import DateFieldEditComponent from "./DateFieldEditComponent";
+import ToggleEditComponent from "./ToggleEditComponent";
+import DropDownSelectComponent from "./DropDownSelectComponent";
 
 /**
  * ProjectTimeline Component - renders the view displayed when the "Timeline"
@@ -82,18 +77,6 @@ const ProjectPhases = ({
     statusMap.find(
       (s) => s.status_name.toLowerCase() === phase_name.toLowerCase()
     );
-
-  // /**
-  //  * Phase table lookup object formatted into the shape that Dropdown expects
-  //  * Ex: { 1: "Potential", 2: "Planned", ...}
-  //  */
-  // const phaseNameLookup = data.moped_phases.reduce(
-  //   (obj, item) =>
-  //     Object.assign(obj, {
-  //       [item.phase_id]: item.phase_name,
-  //     }),
-  //   {}
-  // );
 
   /**
    * Subphase table lookup object formatted into the shape that <MaterialTable>
@@ -145,7 +128,7 @@ const ProjectPhases = ({
    * @returns {Object} Object that will be used in updates to project status
    */
   const getProjectStatusUpdateObject = (mutationPhaseId) => {
-    const newPhaseName = phaseNameLookup[mutationPhaseId];
+    const newPhaseName = phaseNameLookup(data)[mutationPhaseId];
     console.log(newPhaseName);
     const statusMapped = getStatusByPhaseName(newPhaseName);
 
@@ -167,125 +150,13 @@ const ProjectPhases = ({
   };
 
   /**
-   * DateFieldEditComponent - renders a Date type Calendar select
-   * @param {object} props - Values passed through Material Table `editComponent`
-   * @return {JSX.Element}
-   * @constructor
-   */
-  const DateFieldEditComponent = (props) => (
-    <TextField
-      name={props.name}
-      label={props.label}
-      type="date"
-      variant="standard"
-      value={props.value}
-      onChange={(e) => props.onChange(e.target.value)}
-      onKeyDown={(e) => handleKeyEvent(e)}
-      InputLabelProps={{
-        shrink: true,
-      }}
-    />
-  );
-
-  /**
-   * ToggleEditComponent - renders a toggle for True/False edit fields
-   * @param {object} props - Values passed through Material Table `editComponent`
-   * @return {JSX.Element}
-   * @constructor
-   */
-  const ToggleEditComponent = (props) => (
-    <Grid component="label" container alignItems="center" spacing={1}>
-      <Grid item>
-        <Switch
-          checked={props.value ?? false}
-          onChange={(e) => props.onChange(!props.value)}
-          color="primary"
-          name={props.name}
-          inputProps={{ "aria-label": "primary checkbox" }}
-          onKeyDown={(e) => handleKeyEvent(e)}
-        />
-      </Grid>
-    </Grid>
-  );
-
-  /**
-   * DropDownSelectComponent - Renders a drop down menu for MaterialTable
-   * @param {object} props - Values passed through Material Table `editComponent`
-   * @param {string} name - Field name
-   * @return {JSX.Element}
-   * @constructor
-   */
-  const DropDownSelectComponent = (props) => {
-    // If the component name is phase_name, then use phaseNameLookup values, otherwise set as null
-    let lookupValues = props.name === "phase_name" ? phaseNameLookup : null;
-
-    // If lookup values is null, then it is a sub-phase list we need to generate
-    if (lookupValues === null) {
-      // First retrieve the sub-phase id's from moped_phases for that specific row
-      const allowedSubphases = props.data.moped_phases.find(
-        (item) => item?.phase_id === Number(props.rowData?.phase_id ?? 0)
-      )?.moped_subphases;
-
-      // If there are no subphases, hide the drop-down.
-      if (!allowedSubphases || allowedSubphases.length === 0) {
-        return null;
-      }
-
-      // We have a usable array of sub-phase ids, generate lookup values,
-      lookupValues = allowedSubphases.reduce((obj, subphase) => {
-        obj[subphase.subphase_id] = subphase.subphase_name;
-        return obj;
-      }, {});
-    }
-
-    // empty subphases can show up as 0, this removes warning in console
-    lookupValues = { ...lookupValues, 0: "" };
-
-    // Proceed normally and generate the drop-down
-    return (
-      <FormControl>
-        <Select
-          id={props.name}
-          value={props.value ?? ""}
-          style={{ minWidth: "8em" }}
-        >
-          {Object.keys(lookupValues).map((key) => {
-            return (
-              <MenuItem
-                onChange={() => props.onChange(key)}
-                onClick={() => props.onChange(key)}
-                onKeyDown={(e) => handleKeyEvent(e)}
-                value={key}
-                key={key}
-              >
-                {lookupValues[key]}
-              </MenuItem>
-            );
-          })}
-          <MenuItem
-            onChange={() => props.onChange("")}
-            onClick={() => props.onChange("")}
-            onKeyDown={(e) => handleKeyEvent(e)}
-            value=""
-          >
-            -
-          </MenuItem>
-        </Select>
-        {props.name === "phase_name" && (
-          <FormHelperText>Required</FormHelperText>
-        )}
-      </FormControl>
-    );
-  };
-
-  /**
    * Column configuration for <MaterialTable> Phases table
    */
   const phasesColumns = [
     {
       title: "Phase name",
       field: "phase_id",
-      lookup: phaseNameLookup,
+      lookup: phaseNameLookup(data),
       validate: (row) => !!row.phase_id,
       editComponent: (props) => (
         <DropDownSelectComponent {...props} name={"phase_name"} data={data} />
